add: user registration

+++ backend/src/main/java/com/venclima/repository/TideRepository.java
@@ -4,6 +4,9 @@ import com.venclima.model.Tide;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface TideRepository extends JpaRepository<Tide, Long> {
+public interface TideRepository extends JpaRepository<Tide, Integer> {
+    //List<Tide> findByStation_Id(int stationId);
 }
